+++ eslint.config.js
@@ -1,7 +1,7 @@
-import antfu from '@antfu/eslint-config';
+import antfu from '@antfu/eslint-config'
 
 export default antfu({
-	formatters: true,
+	formatters: false,
 
 	stylistic: {
 		indent: 'tab',
@@ -17,4 +17,4 @@ export default antfu({
 	},
 
 	ignores: [],
-});
+})